+++ apps/react-lib-dev/src/app/app.tsx
@@ -9,7 +9,7 @@ import {
   Stepper,
   TextInput,
   Datepicker,
-  Alert
+  Alert,
 } from '@sebgroup/green-react'
 import { useState } from 'react'
 import { IValidator } from '@sebgroup/extract'
@@ -45,7 +45,9 @@ export function App() {
     <>
       <div className="use-green">
         <Navbar title="Green React Dev" />
-        <Alert type={"warning"} isCloseable={false}>This is my alert</Alert>
+        <Alert type={'warning'} isCloseable={false}>
+          This is my alert
+        </Alert>
         <div className="container">
           <div className="row">
             <div className="col">
@@ -88,7 +90,7 @@ export function App() {
                   }}
                 >
                   <RadioGroup
-                    title="Radio Group"
+                    label="Radio Group"
                     onChange={(value) => console.log(value)}
                   >
                     <RadioButton label={'Fusilli'} value={'fusilli'} />