+++ libs/react/src/lib/select/select.tsx
@@ -6,13 +6,10 @@ import {
   validateClassName,
 } from '@sebgroup/extract'
 import { ChevronDown } from '../icons'
+import { SelectorAttributesProps, LabelProps } from '../../types'
 import { FormItem } from '../formItem'
 
-export interface SelectProps {
-  id?: string
-  className?: string
-  label?: string
-  labelInformation?: string
+export interface SelectProps extends SelectorAttributesProps, LabelProps {
   children: ReactNode
   validator?: IValidator
   onChange?: (event: React.ChangeEvent<HTMLSelectElement>) => void
@@ -24,7 +21,7 @@ export interface SelectProps {
   expandableInfoButtonLabel?: string
 }
 
-export interface OptionProps {
+interface OptionProps {
   children: string
   value: string | number
   disabled?: boolean
@@ -38,7 +35,7 @@ export interface OptionProps {
   selected?: boolean
 }
 
-export interface OptionGroupProps {
+interface OptionGroupProps {
   label: string
   children: React.ReactElement<OptionProps> | React.ReactElement<OptionProps>[]
   disabled?: boolean
@@ -117,5 +114,4 @@ export const OptionGroup = ({
     </optgroup>
   )
 }
-
 export default { Select, Option, OptionGroup }