+++ libs/react/src/types/props/index.ts
@@ -4,3 +4,14 @@ export interface IconProps {
   fill?: string
   'aria-hidden'?: boolean
 }
+
+export interface SelectorAttributesProps {
+  id?: string
+  className?: string
+  testId?: string
+}
+
+export interface LabelProps {
+  label?: string
+  labelInformation?: string
+}