+++ libs/angular-charts/CHANGELOG.md
@@ -2,6 +2,38 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.0.0](https://github.com/sebgroup/green/compare/@sebgroup/green-angular-charts@0.0.1...@sebgroup/green-angular-charts@1.0.0) (2023-02-20)
+
+
+### chore
+
+* 🤖 bumping react and angular charts version ([31c93e1](https://github.com/sebgroup/green/commit/31c93e18f37119ff033868538c963311ee50db44))
+
+
+### BREAKING CHANGES
+
+* 🧨 bumping to 1.0.0
+
+
+
+# [1.0.0](https://github.com/sebgroup/green/compare/@sebgroup/green-angular-charts@0.0.1...@sebgroup/green-angular-charts@1.0.0) (2023-02-20)
+
+
+### chore
+
+* 🤖 bumping react and angular charts version ([31c93e1](https://github.com/sebgroup/green/commit/31c93e18f37119ff033868538c963311ee50db44))
+
+
+### BREAKING CHANGES
+
+* 🧨 bumping to 1.0.0
+
+
+
+# [1.0.0](https://github.com/sebgroup/green/compare/@sebgroup/green-angular-charts@0.0.1...@sebgroup/green-angular-charts@1.0.0) (2023-02-17)
+
+
+
 ## [0.0.1](https://github.com/sebgroup/green/compare/@sebgroup/green-angular-charts@1.0.0-beta.45...@sebgroup/green-angular-charts@0.0.1) (2023-01-27)
 
 ### Dependency Updates