+++ libs/charts/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.0.1](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0...@sebgroup/green-charts@1.0.1) (2023-02-20)
+
+
+
+# [1.0.0](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@0.0.1...@sebgroup/green-charts@1.0.0) (2023-02-17)
+
+
+
 ## [0.0.1](https://github.com/sebgroup/green/compare/@sebgroup/green-charts@1.0.0-beta.44...@sebgroup/green-charts@0.0.1) (2023-01-27)
 
 ### Dependency Updates