+++ libs/extract/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.0.0](https://github.com/sebgroup/green/compare/@sebgroup/extract@0.0.2...@sebgroup/extract@1.0.0) (2023-02-17)
+
+
+
 ## [0.0.2](https://github.com/sebgroup/green/compare/@sebgroup/extract@0.0.1...@sebgroup/extract@0.0.2) (2023-02-02)